@@ -1,17 +1,29 @@
 module go.breu.io/quantm
 
-go 1.20
+go 1.21.0
+
+toolchain go1.21.6
 
 require (
+	cloud.google.com/go/artifactregistry v1.14.7
+	cloud.google.com/go/iam v1.1.5
+	cloud.google.com/go/logging v1.9.0
+	cloud.google.com/go/run v1.3.3
+	github.com/GoogleCloudPlatform/opentelemetry-operations-go/exporter/trace v1.21.0
+	github.com/GoogleCloudPlatform/opentelemetry-operations-go/propagator v0.45.0
 	github.com/Guilospanck/gocqlxmock v1.0.1
 	github.com/Guilospanck/igocqlx v1.0.0
 	github.com/Pallinder/go-randomdata v1.2.0
 	github.com/avast/retry-go/v4 v4.5.1
 	github.com/bradleyfalzon/ghinstallation/v2 v2.9.0
-	github.com/deepmap/oapi-codegen v1.13.3
+	github.com/charmbracelet/bubbles v0.17.1
+	github.com/charmbracelet/bubbletea v0.25.0
+	github.com/charmbracelet/lipgloss v0.9.1
+	github.com/deepmap/oapi-codegen/v2 v2.0.0
 	github.com/docker/docker v24.0.7+incompatible
 	github.com/fatih/color v1.16.0
 	github.com/gabriel-vasile/mimetype v1.4.2 // indirect
+	github.com/go-git/go-git/v5 v5.11.0
 	github.com/go-playground/validator/v10 v10.17.0
 	github.com/gocql/gocql v1.6.0
 	github.com/golang-jwt/jwt/v4 v4.5.0
