@@ -19,7 +19,8 @@
 package auth
 
 import (
-	_ "github.com/deepmap/oapi-codegen/pkg/codegen" // Required for code generation
+	_ "github.com/deepmap/oapi-codegen/v2/pkg/codegen" // Required for code generation
+	_ "gopkg.in/yaml.v2"
 )
 
-//go:generate go run github.com/deepmap/oapi-codegen/cmd/oapi-codegen --config openapi.codegen.yaml openapi.spec.yaml
+//go:generate go run github.com/deepmap/oapi-codegen/v2/cmd/oapi-codegen --config openapi.codegen.yaml openapi.spec.yaml
