@@ -1,6 +1,6 @@
 // Package auth provides primitives to interact with the openapi HTTP API.
 //
-// Code generated by github.com/deepmap/oapi-codegen, a modified copy of github.com/deepmap/oapi-codegen.
+// Code generated by github.com/deepmap/oapi-codegen, a modified copy of github.com/deepmap/oapi-codegen/v2.
 // It was modified to add support for the following features:
 //  - Support for custom templates by filename.
 //  - Supporting x-breu-entity in the schema to generate a struct for the entity.
@@ -21,10 +21,10 @@ import (
 	"time"
 
 	itable "github.com/Guilospanck/igocqlx/table"
-	"github.com/deepmap/oapi-codegen/pkg/runtime"
-	openapi_types "github.com/deepmap/oapi-codegen/pkg/types"
 	"github.com/gocql/gocql"
 	"github.com/labstack/echo/v4"
+	"github.com/oapi-codegen/runtime"
+	openapi_types "github.com/oapi-codegen/runtime/types"
 	"github.com/scylladb/gocqlx/v2/table"
 	"go.breu.io/quantm/internal/shared"
 	externalRef0 "go.breu.io/quantm/internal/shared"
@@ -85,11 +85,11 @@ type RegisterationResponse struct {
 
 // Team defines model for Team.
 type Team struct {
-	CreatedAt time.Time  `cql:"created_at" json:"created_at"`
-	ID        gocql.UUID `cql:"id" json:"id"`
-	Name      string     `cql:"name" json:"name" validate:"required"`
-	Slug      string     `cql:"slug" json:"slug"`
-	UpdatedAt time.Time  `cql:"updated_at" json:"updated_at"`
+	CreatedAt time.Time  `json:"created_at"`
+	ID        gocql.UUID `json:"id"`
+	Name      string     `json:"name" validate:"required"`
+	Slug      string     `json:"slug"`
+	UpdatedAt time.Time  `json:"updated_at"`
 }
 
 var (
@@ -116,11 +116,11 @@ type TeamList struct {
 
 // TeamUser defines model for TeamUser.
 type TeamUser struct {
-	CreatedAt time.Time  `cql:"created_at" json:"created_at"`
-	ID        gocql.UUID `cql:"id" json:"id"`
-	TeamID    gocql.UUID `cql:"team_id" json:"team_id"`
-	UpdatedAt time.Time  `cql:"updated_at" json:"updated_at"`
-	UserID    gocql.UUID `cql:"user_id" json:"user_id"`
+	CreatedAt time.Time  `json:"created_at"`
+	ID        gocql.UUID `json:"id"`
+	TeamID    gocql.UUID `json:"team_id"`
+	UpdatedAt time.Time  `json:"updated_at"`
+	UserID    gocql.UUID `json:"user_id"`
 }
 
 var (
@@ -148,16 +148,16 @@ type TokenResponse struct {
 
 // User defines model for User.
 type User struct {
-	CreatedAt  time.Time  `cql:"created_at" json:"created_at"`
-	Email      string     `cql:"email" json:"email" validate:"email,required,db_unique"`
-	FirstName  string     `cql:"first_name" json:"first_name"`
-	ID         gocql.UUID `cql:"id" json:"id"`
-	IsActive   bool       `cql:"is_active" json:"is_active"`
-	IsVerified bool       `cql:"is_verified" json:"is_verified"`
-	LastName   string     `cql:"last_name" json:"last_name"`
-	Password   string     `cql:"password" json:"-"`
-	TeamID     gocql.UUID `cql:"team_id" json:"team_id"`
-	UpdatedAt  time.Time  `cql:"updated_at" json:"updated_at"`
+	CreatedAt  time.Time  `json:"created_at"`
+	Email      string     `json:"email" validate:"email,required,db_unique"`
+	FirstName  string     `json:"first_name"`
+	ID         gocql.UUID `json:"id"`
+	IsActive   bool       `json:"is_active"`
+	IsVerified bool       `json:"is_verified"`
+	LastName   string     `json:"last_name"`
+	Password   string     `json:"-"`
+	TeamID     gocql.UUID `json:"team_id"`
+	UpdatedAt  time.Time  `json:"updated_at"`
 }
 
 var (
