@@ -1,6 +1,6 @@
 // Package core provides primitives to interact with the openapi HTTP API.
 //
-// Code generated by github.com/deepmap/oapi-codegen, a modified copy of github.com/deepmap/oapi-codegen.
+// Code generated by github.com/deepmap/oapi-codegen, a modified copy of github.com/deepmap/oapi-codegen/v2.
 // It was modified to add support for the following features:
 //  - Support for custom templates by filename.
 //  - Supporting x-breu-entity in the schema to generate a struct for the entity.
@@ -22,9 +22,9 @@ import (
 	"time"
 
 	itable "github.com/Guilospanck/igocqlx/table"
-	"github.com/deepmap/oapi-codegen/pkg/runtime"
 	"github.com/gocql/gocql"
 	"github.com/labstack/echo/v4"
+	"github.com/oapi-codegen/runtime"
 	"github.com/scylladb/gocqlx/v2/table"
 	"go.breu.io/quantm/internal/shared"
 	externalRef0 "go.breu.io/quantm/internal/shared"
@@ -180,16 +180,16 @@ type BluePrintRegions struct {
 
 // Blueprint Blueprint contains a collection of Workload & Resource to define one single release
 type Blueprint struct {
-	CreatedAt      time.Time  `cql:"created_at" json:"created_at"`
-	ID             gocql.UUID `cql:"id" json:"id"`
-	Name           string     `cql:"name" json:"name"`
-	ProviderConfig string     `cql:"provider_config" json:"provider_config"`
+	CreatedAt      time.Time  `json:"created_at"`
+	ID             gocql.UUID `json:"id"`
+	Name           string     `json:"name"`
+	ProviderConfig string     `json:"provider_config"`
 
 	// Regions BluePrintRegions sets the cloud regions where a blueprint can be deployed
-	Regions       BluePrintRegions `cql:"regions" json:"regions"`
-	RolloutBudget string           `cql:"rollout_budget" json:"rollout_budget"`
-	StackID       gocql.UUID       `cql:"stack_id" json:"stack_id"`
-	UpdatedAt     time.Time        `cql:"updated_at" json:"updated_at"`
+	Regions       BluePrintRegions `json:"regions"`
+	RolloutBudget string           `json:"rollout_budget"`
+	StackID       gocql.UUID       `json:"stack_id"`
+	UpdatedAt     time.Time        `json:"updated_at"`
 }
 
 var (
@@ -228,15 +228,15 @@ type Driver string
 
 // Repo defines model for Repo.
 type Repo struct {
-	CreatedAt     time.Time    `cql:"created_at" json:"created_at"`
-	DefaultBranch string       `cql:"default_branch" json:"default_branch"`
-	ID            gocql.UUID   `cql:"id" json:"id"`
-	IsMonorepo    bool         `cql:"is_monorepo" json:"is_monorepo"`
-	Name          string       `cql:"name" json:"name"`
-	Provider      RepoProvider `cql:"provider" json:"provider"`
-	ProviderID    string       `cql:"provider_id" json:"provider_id"`
-	StackID       gocql.UUID   `cql:"stack_id" json:"stack_id"`
-	UpdatedAt     time.Time    `cql:"updated_at" json:"updated_at"`
+	CreatedAt     time.Time    `json:"created_at"`
+	DefaultBranch string       `json:"default_branch"`
+	ID            gocql.UUID   `json:"id"`
+	IsMonorepo    bool         `json:"is_monorepo"`
+	Name          string       `json:"name"`
+	Provider      RepoProvider `json:"provider"`
+	ProviderID    string       `json:"provider_id"`
+	StackID       gocql.UUID   `json:"stack_id"`
+	UpdatedAt     time.Time    `json:"updated_at"`
 }
 
 var (
@@ -275,19 +275,19 @@ type RepoProvider string
 // Resource Resource defines the cloud provider resources for the app e.g. s3, sqs, etc
 type Resource struct {
 	// Config resource configruation e.g properties, output environment variables etc
-	Config    string    `cql:"config" json:"config"`
-	CreatedAt time.Time `cql:"created_at" json:"created_at"`
+	Config    string    `json:"config"`
+	CreatedAt time.Time `json:"created_at"`
 
 	// Driver gke, cloudrun, pubsub, s3, sqs, sns, dynamodb, postgres, mysql etc
-	Driver      Driver     `cql:"driver" json:"driver"`
-	ID          gocql.UUID `cql:"id" json:"id"`
-	IsImmutable bool       `cql:"is_immutable" json:"is_immutable"`
-	Name        string     `cql:"name" json:"name"`
+	Driver      Driver     `json:"driver"`
+	ID          gocql.UUID `json:"id"`
+	IsImmutable bool       `json:"is_immutable"`
+	Name        string     `json:"name"`
 
 	// Provider aws, gcp, azure
-	Provider  CloudProvider `cql:"provider" json:"provider"`
-	StackID   gocql.UUID    `cql:"stack_id" json:"stack_id"`
-	UpdatedAt time.Time     `cql:"updated_at" json:"updated_at"`
+	Provider  CloudProvider `json:"provider"`
+	StackID   gocql.UUID    `json:"stack_id"`
+	UpdatedAt time.Time     `json:"updated_at"`
 }
 
 var (
@@ -324,13 +324,13 @@ type ResourceListResponse = []Resource
 
 // Stack defines model for Stack.
 type Stack struct {
-	Config    StackConfig `cql:"config" json:"config"`
-	CreatedAt time.Time   `cql:"created_at" json:"created_at"`
-	ID        gocql.UUID  `cql:"id" json:"id"`
-	Name      string      `cql:"name" json:"name" validate:"required"`
-	Slug      string      `cql:"slug" json:"slug"`
-	TeamID    gocql.UUID  `cql:"team_id" json:"team_id"`
-	UpdatedAt time.Time   `cql:"updated_at" json:"updated_at"`
+	Config    StackConfig `json:"config"`
+	CreatedAt time.Time   `json:"created_at"`
+	ID        gocql.UUID  `json:"id"`
+	Name      string      `json:"name" validate:"required"`
+	Slug      string      `json:"slug"`
+	TeamID    gocql.UUID  `json:"team_id"`
+	UpdatedAt time.Time   `json:"updated_at"`
 }
 
 var (
@@ -365,21 +365,21 @@ type StackListResponse = []Stack
 // Workload Workload defines a workload for the app
 type Workload struct {
 	// Builder json with keys: buildpack, dockerfile, script, external
-	Builder string `cql:"builder" json:"builder"`
+	Builder string `json:"builder"`
 
 	// Container json with keys: image, command, environment, dependencies
-	Container string     `cql:"container" json:"container"`
-	CreatedAt time.Time  `cql:"created_at" json:"created_at"`
-	ID        gocql.UUID `cql:"id" json:"id"`
+	Container string     `json:"container"`
+	CreatedAt time.Time  `json:"created_at"`
+	ID        gocql.UUID `json:"id"`
 
 	// Kind Default, worker, job, cronjob
-	Kind       string     `cql:"kind" json:"kind"`
-	Name       string     `cql:"name" json:"name"`
-	RepoID     gocql.UUID `cql:"repo_id" json:"repo_id"`
-	RepoPath   string     `cql:"repo_path" json:"repo_path"`
-	ResourceID gocql.UUID `cql:"resource_id" json:"resource_id"`
-	StackID    gocql.UUID `cql:"stack_id" json:"stack_id"`
-	UpdatedAt  time.Time  `cql:"updated_at" json:"updated_at"`
+	Kind       string     `json:"kind"`
+	Name       string     `json:"name"`
+	RepoID     gocql.UUID `json:"repo_id"`
+	RepoPath   string     `json:"repo_path"`
+	ResourceID gocql.UUID `json:"resource_id"`
+	StackID    gocql.UUID `json:"stack_id"`
+	UpdatedAt  time.Time  `json:"updated_at"`
 }
 
 var (
