@@ -19,8 +19,8 @@
 package shared
 
 import (
-	_ "github.com/deepmap/oapi-codegen/pkg/codegen" // Required for code generation
+	_ "github.com/deepmap/oapi-codegen/v2/pkg/codegen" // Required for code generation
 	_ "gopkg.in/yaml.v2"
 )
 
-//go:generate go run github.com/deepmap/oapi-codegen/cmd/oapi-codegen -config openapi.codegen.yaml -package shared -generate types,skip-prune,client -o types.gen.go openapi.spec.yaml
+//go:generate go run github.com/deepmap/oapi-codegen/v2/cmd/oapi-codegen -config openapi.codegen.yaml -package shared -generate types,skip-prune,client -o types.gen.go openapi.spec.yaml
