@@ -15,10 +15,57 @@
 // CONSEQUENTIAL, SPECIAL, INCIDENTAL, INDIRECT, OR DIRECT DAMAGES, HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
 // ARISING OUT OF THIS AGREEMENT. THE FOREGOING SHALL APPLY TO THE EXTENT PERMITTED BY APPLICABLE LAW.
 
-package core
+package logger
 
 import (
-	_ "github.com/deepmap/oapi-codegen/pkg/codegen" // Required for code generation
+	"log/slog"
+	"os"
+	"sync"
+
+	"cloud.google.com/go/compute/metadata"
+	"cloud.google.com/go/logging"
 )
 
-//go:generate go run github.com/deepmap/oapi-codegen/cmd/oapi-codegen --config _schema/gen.yaml _schema/openapi.yaml
+type (
+	Level2Severity map[slog.Level]logging.Severity
+)
+
+var (
+	project     string
+	projectOnce sync.Once
+)
+
+const (
+	LevelDefault   = 0
+	LevelDebug     = 100
+	LevelInfo      = 200
+	LevelNotice    = 300
+	LevelWarning   = 400
+	LevelError     = 500
+	LevelCritical  = 600
+	LevelAlert     = 700
+	LevelEmergency = 800
+)
+
+func SeverityFromLevel(level slog.Level) logging.Severity {
+	translator := Level2Severity{
+		slog.LevelDebug: logging.Debug,
+		slog.LevelInfo:  logging.Info,
+		slog.LevelWarn:  logging.Warning,
+		slog.LevelError: logging.Error,
+	}
+
+	return translator[level]
+}
+
+func GoogleCloudProject() string {
+	projectOnce.Do(func() {
+		if metadata.OnGCE() {
+			project, _ = metadata.ProjectID()
+		} else {
+			project = os.Getenv("GOOGLE_CLOUD_PROJECT") // This is required for local development.
+		}
+	})
+
+	return project
+}
