@@ -1,6 +1,6 @@
 // Package shared provides primitives to interact with the openapi HTTP API.
 //
-// Code generated by github.com/deepmap/oapi-codegen version v1.13.3 DO NOT EDIT.
+// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.0.0 DO NOT EDIT.
 package shared
 
 import (
